Refactor: move 'FieldPath' / related helpers to a new 'field_path' module.
