+++ firestore/docs/field_path.rst
@@ -1,0 +1,7 @@
+Field Paths
+~~~~~~~~~~~
+
+.. automodule:: google.cloud.firestore_v1beta1.field_path
+  :members:
+  :show-inheritance:
+
