+++ firestore/google/cloud/firestore_v1beta1/_helpers.py
@@ -14,13 +14,7 @@
 
 """Common helpers shared across Google Cloud Firestore modules."""
 
-try:
-    from collections import abc as collections_abc
-except ImportError:  # Python 2.7
-    import collections as collections_abc
-
 import datetime
-import re
 
 from google.protobuf import struct_pb2
 from google.type import latlng_pb2
@@ -32,6 +26,8 @@ from google.cloud._helpers import _datetime_to_pb_timestamp
 from google.cloud._helpers import _pb_timestamp_to_datetime
 from google.cloud.firestore_v1beta1 import transforms
 from google.cloud.firestore_v1beta1 import types
+from google.cloud.firestore_v1beta1.field_path import FieldPath
+from google.cloud.firestore_v1beta1.field_path import parse_field_path
 from google.cloud.firestore_v1beta1.gapic import enums
 from google.cloud.firestore_v1beta1.proto import common_pb2
 from google.cloud.firestore_v1beta1.proto import document_pb2
@@ -39,12 +35,6 @@ from google.cloud.firestore_v1beta1.proto import write_pb2
 
 
 BAD_PATH_TEMPLATE = "A path element must be a string. Received {}, which is a {}."
-FIELD_PATH_MISSING_TOP = "{!r} is not contained in the data"
-FIELD_PATH_MISSING_KEY = "{!r} is not contained in the data for the key {!r}"
-FIELD_PATH_WRONG_TYPE = (
-    "The data at {!r} is not a dictionary, so it cannot contain the key {!r}"
-)
-FIELD_PATH_DELIMITER = "."
 DOCUMENT_PATH_DELIMITER = "/"
 INACTIVE_TXN = "Transaction not in progress, cannot be used in API requests."
 READ_AFTER_WRITE_ERROR = "Attempted read after write in a transaction."
@@ -109,111 +99,6 @@ class GeoPoint(object):
             return NotImplemented
         else:
             return not equality_val
-
-
-class FieldPath(object):
-    """ Field Path object for client use.
-
-    Args:
-        parts: (one or more strings)
-            Indicating path of the key to be used.
-    """
-
-    def __init__(self, *parts):
-        for part in parts:
-            if not isinstance(part, six.string_types) or not part:
-                error = "One or more components is not a string or is empty."
-                raise ValueError(error)
-        self.parts = tuple(parts)
-
-    @staticmethod
-    def from_string(string):
-        """ Creates a FieldPath from a unicode string representation.
-
-        This method splits on the character `.` and disallows the
-        characters `~*/[]`. To create a FieldPath whose components have
-        those characters, call the constructor.
-
-        Args:
-            :type string: str
-            :param string: A unicode string which cannot contain
-                           `~*/[]` characters, cannot exceed 1500 bytes,
-                           and cannot be empty.
-
-        Returns:
-            A :class: `FieldPath` instance with the string split on "."
-            as arguments to `FieldPath`.
-        """
-        # XXX this should just handle things with the invalid chars
-        invalid_characters = "~*/[]"
-        for invalid_character in invalid_characters:
-            if invalid_character in string:
-                raise ValueError("Invalid characters in string.")
-        string = string.split(".")
-        return FieldPath(*string)
-
-    def __repr__(self):
-        paths = ""
-        for part in self.parts:
-            paths += "'" + part + "',"
-        paths = paths[:-1]
-        return "FieldPath({})".format(paths)
-
-    def __hash__(self):
-        return hash(self.to_api_repr())
-
-    def __eq__(self, other):
-        if isinstance(other, FieldPath):
-            return self.parts == other.parts
-        return NotImplemented
-
-    def __lt__(self, other):
-        if isinstance(other, FieldPath):
-            return self.parts < other.parts
-        return NotImplemented
-
-    def __add__(self, other):
-        """Adds `other` field path to end of this field path.
-
-        Args:
-            other (~google.cloud.firestore_v1beta1._helpers.FieldPath, str):
-                The field path to add to the end of this `FieldPath`.
-        """
-        if isinstance(other, FieldPath):
-            parts = self.parts + other.parts
-            return FieldPath(*parts)
-        elif isinstance(other, six.string_types):
-            parts = self.parts + FieldPath.from_string(other).parts
-            return FieldPath(*parts)
-        else:
-            return NotImplemented
-
-    def eq_or_parent(self, other):
-        return self.parts[: len(other.parts)] == other.parts[: len(self.parts)]
-
-    def to_api_repr(self):
-        """ Returns quoted string representation of the FieldPath
-
-        Returns: :rtype: str
-            Quoted string representation of the path stored
-            within this FieldPath conforming to the Firestore API
-            specification
-        """
-        return get_field_path(self.parts)
-
-    def lineage(self):
-        """Return field paths for all parents.
-
-        Returns: set(FieldPath)
-        """
-        parts = self.parts[:-1]
-        result = set()
-
-        while parts:
-            result.add(FieldPath(*parts))
-            parts = parts[:-1]
-
-        return result
 
 
 def verify_path(path, is_collection):
@@ -427,217 +312,6 @@ def decode_dict(value_fields, client):
     return {
         key: decode_value(value, client) for key, value in six.iteritems(value_fields)
     }
-
-
-SIMPLE_FIELD_NAME = re.compile("^[_a-zA-Z][_a-zA-Z0-9]*$")
-
-
-def get_field_path(field_names):
-    """Create a **field path** from a list of nested field names.
-
-    A **field path** is a ``.``-delimited concatenation of the field
-    names. It is used to represent a nested field. For example,
-    in the data
-
-    .. code-block: python
-
-       data = {
-          'aa': {
-              'bb': {
-                  'cc': 10,
-              },
-          },
-       }
-
-    the field path ``'aa.bb.cc'`` represents that data stored in
-    ``data['aa']['bb']['cc']``.
-
-    Args:
-        field_names (Iterable[str, ...]): The list of field names.
-
-    Returns:
-        str: The ``.``-delimited field path.
-    """
-    result = []
-
-    for field_name in field_names:
-        match = SIMPLE_FIELD_NAME.match(field_name)
-        if match and match.group(0) == field_name:
-            result.append(field_name)
-        else:
-            replaced = field_name.replace("\\", "\\\\").replace("`", "\\`")
-            result.append("`" + replaced + "`")
-
-    return FIELD_PATH_DELIMITER.join(result)
-
-
-PATH_ELEMENT_TOKENS = [
-    ("SIMPLE", r"[_a-zA-Z][_a-zA-Z0-9]*"),  # unquoted elements
-    ("QUOTED", r"`(?:\\`|[^`])*?`"),  # quoted elements, unquoted
-    ("DOT", r"\."),  # separator
-]
-TOKENS_PATTERN = "|".join("(?P<{}>{})".format(*pair) for pair in PATH_ELEMENT_TOKENS)
-TOKENS_REGEX = re.compile(TOKENS_PATTERN)
-
-
-def _tokenize_field_path(path):
-    """Lex a field path into tokens (including dots).
-
-    Args:
-        path (str): field path to be lexed.
-    Returns:
-        List(str): tokens
-    """
-    pos = 0
-    get_token = TOKENS_REGEX.match
-    match = get_token(path)
-    while match is not None:
-        type_ = match.lastgroup
-        value = match.group(type_)
-        yield value
-        pos = match.end()
-        match = get_token(path, pos)
-
-
-def split_field_path(path):
-    """Split a field path into valid elements (without dots).
-
-    Args:
-        path (str): field path to be lexed.
-    Returns:
-        List(str): tokens
-    Raises:
-        ValueError: if the path does not match the elements-interspersed-
-                    with-dots pattern.
-    """
-    if not path:
-        return []
-
-    elements = []
-    want_dot = False
-
-    for element in _tokenize_field_path(path):
-        if want_dot:
-            if element != ".":
-                raise ValueError("Invalid path: {}".format(path))
-            else:
-                want_dot = False
-        else:
-            if element == ".":
-                raise ValueError("Invalid path: {}".format(path))
-            elements.append(element)
-            want_dot = True
-
-    if not want_dot or not elements:
-        raise ValueError("Invalid path: {}".format(path))
-
-    return elements
-
-
-def parse_field_path(api_repr):
-    """Parse a **field path** from into a list of nested field names.
-
-    See :func:`field_path` for more on **field paths**.
-
-    Args:
-        api_repr (str):
-            The unique Firestore api representation which consists of
-            either simple or UTF-8 field names. It cannot exceed
-            1500 bytes, and cannot be empty. Simple field names match
-            `'^[_a-zA-Z][_a-zA-Z0-9]*$'`. All other field names are
-            escaped with ```.
-
-    Returns:
-        List[str, ...]: The list of field names in the field path.
-    """
-    # code dredged back up from
-    # https://github.com/googleapis/google-cloud-python/pull/5109/files
-    field_names = []
-    for field_name in split_field_path(api_repr):
-        # non-simple field name
-        if field_name[0] == "`" and field_name[-1] == "`":
-            field_name = field_name[1:-1]
-            field_name = field_name.replace("\\`", "`")
-            field_name = field_name.replace("\\\\", "\\")
-        field_names.append(field_name)
-    return field_names
-
-
-def get_nested_value(field_path, data):
-    """Get a (potentially nested) value from a dictionary.
-
-    If the data is nested, for example:
-
-    .. code-block:: python
-
-       >>> data
-       {
-           'top1': {
-               'middle2': {
-                   'bottom3': 20,
-                   'bottom4': 22,
-               },
-               'middle5': True,
-           },
-           'top6': b'\x00\x01 foo',
-       }
-
-    a **field path** can be used to access the nested data. For
-    example:
-
-    .. code-block:: python
-
-       >>> get_nested_value('top1', data)
-       {
-           'middle2': {
-               'bottom3': 20,
-               'bottom4': 22,
-           },
-           'middle5': True,
-       }
-       >>> get_nested_value('top1.middle2', data)
-       {
-           'bottom3': 20,
-           'bottom4': 22,
-       }
-       >>> get_nested_value('top1.middle2.bottom3', data)
-       20
-
-    See :meth:`~.firestore_v1beta1.client.Client.field_path` for
-    more information on **field paths**.
-
-    Args:
-        field_path (str): A field path (``.``-delimited list of
-            field names).
-        data (Dict[str, Any]): The (possibly nested) data.
-
-    Returns:
-        Any: (A copy of) the value stored for the ``field_path``.
-
-    Raises:
-        KeyError: If the ``field_path`` does not match nested data.
-    """
-    field_names = parse_field_path(field_path)
-
-    nested_data = data
-    for index, field_name in enumerate(field_names):
-        if isinstance(nested_data, collections_abc.Mapping):
-            if field_name in nested_data:
-                nested_data = nested_data[field_name]
-            else:
-                if index == 0:
-                    msg = FIELD_PATH_MISSING_TOP.format(field_name)
-                    raise KeyError(msg)
-                else:
-                    partial = get_field_path(field_names[:index])
-                    msg = FIELD_PATH_MISSING_KEY.format(field_name, partial)
-                    raise KeyError(msg)
-        else:
-            partial = get_field_path(field_names[:index])
-            msg = FIELD_PATH_WRONG_TYPE.format(partial, field_name)
-            raise KeyError(msg)
-
-    return nested_data
 
 
 def get_doc_id(document_pb, expected_prefix):
